pyproject.toml
[project]
name = "dismech-jax"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Ryan Chaiyakul", email = "ryanchaiyakul@gmail.com" }]
requires-python = ">=3.13"
dependencies = ["jax>=0.7.0"]

[project.optional-dependencies]
cuda = ["jax[cuda12]>=0.7.0"]
tpu = ["jax[tpu]>=0.7.0"]

[tool.mypy]
files = ["src/**/*.py", "tests/**/*.py"]
strict = true
allow_redefinition = true
pretty = true

[tool.ruff]
include = ["pyproject.toml", "src/**/*.py", "tests/**/*.py"]

[build-system]
requires = ["uv_build>=0.8.8,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "chex>=0.1.90",
    "ipykernel>=6.30.1",
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "pytest-benchmark>=5.1.0",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.8",
    "scipy-stubs>=1.16.1.0",
    "sphinx>=8.2.3",
    "tensorflow>=2.20.0rc0",
]
